Quote deletion improvement
Improvement is subjective. I do not think my solutions are good performance wise. Corrected the issue of qids (quote ids) not matching up with the appropriate quote when deleting it and backend now checks if a quote was successfully deleted or not

+++ routes/quotes.js
@@ -32,14 +32,18 @@ router.route('/quotes')
 		// TODO: Instead of storing the quotes in an array and rendering it, store them in the session
 		// 		 so we can render the page without having to wait for the firestore.
 		// TODO: ignore improperly formatted quotes
+		// TODO: pagination
 		const displayQuotes = [];	// array of quote objects
 		allUserSnaps.forEach((userSnap) => {	// loops through all documents
 			const { username, quotes } = userSnap.data();
 			displayQuotes.push(...generateQuoteObjects(quotes, username));
 		});
 		displayQuotes.sort(require('../utils/models').Quote.sortDatesDescending);
-		console.log(displayQuotes.length)
-		return res.status(200).render(quotesPagePath, { quotes: JSON.stringify(displayQuotes) });
+		const currUserQuotes = displayQuotes.filter((quote) => quote.author === auth.currentUser.displayName);
+		return res.status(200).render(quotesPagePath, {
+			quotes: displayQuotes,
+			currUserQuotes: JSON.stringify(currUserQuotes)
+		});
 	})
 	// Create quote
 	.post(async (req, res) => {
@@ -57,10 +61,6 @@ router.route('/quotes')
 				determineError(err);
 			}
 		}
-		// Typically, the alert messages are hangled on the backend, but
-		// I did it on the frontend this time cause of the modal.
-		// The if-statement exists in case someone submits a POST request
-		// through an external application (i.e. Postman).
 		// Refresh the page in case others have posted new quotes.
 		res.status(201).redirect('/quotes');	// NOTE: setting status like this with .redirect doesn't matter
 	});
@@ -69,23 +69,41 @@ router.route('/quotes')
 router.route('/quotes/:qid')
 	// Votes, quote body
 	.post(async (req, res) => {
-		// TODO: cannot vote on things that currently have a vote
-		// TODO: cannot vote on their own quotes
+		// TODO: make it so users cannot upvote if they already upvote and same with downvotes
+		// TODO: make it so users cannot vote on their own quotes
 		const { vote } = req.body;
 	})
 	// Quote
 	.delete(async (req, res) => {
 		// Security rules will prevent users from deleting quotes that aren't theirs
+		// TODO: optimize this (spending two reads just to check if a single quote was deleted)
 		const { qid } = req.params;
 		res.setHeader('Content-Type', 'text/plain');
 		try {
-			await usersColl.doc(auth.currentUser.uid).update({
-				[`quotes.${qid}`]: fv.delete()
-			});
+			// check if quote exists
+			let quote = (await usersColl.doc(auth.currentUser.uid).get()).data().quotes[qid];
+			if (quote) {
+				// delete quote
+				await usersColl.doc(auth.currentUser.uid).update({
+					[`quotes.${qid}`]: fv.delete()
+				});
+				// check if it was deleted
+				quote = (await usersColl.doc(auth.currentUser.uid).get()).data().quotes[qid];
+				if (!quote) {
+					// if it doesn't exist anymore, it was successfully deleted
+					return res.sendStatus(204);
+				} else {
+					// the quote was not deleted
+					return res.status(500).send('Could not delete quote.');
+				}
+			} else {
+				// a quote is not associated with the qid
+				return res.status(200).send('Quote does not exist.');
+			}
 		} catch (err) {
-			return res.status(403).send('Cannot delete quotes that are not your own.');
+			determineError(err);	// TODO: delete this (temporary)
+			return res.status(403).send('Can\'t delete quotes that aren\'t yours.');
 		}
-		return res.sendStatus(200);
 	});
 
 module.exports = router;
